refactor(Popover): sync popover with menu and select components
- alignment with the code in the other popover-like components
- some light changes to component structure
- cleaning up the structure of the components directory to merge
  implied sub-components together
