--- src/components/PopoverButton/index.ts
@@ -1,1 +1,0 @@
-export { PopoverButton as default } from './PopoverButton';
