--- src/components/PopoverContent/index.ts
@@ -1,1 +1,0 @@
-export { PopoverContent as default } from './PopoverContent';
